Send PRs to forked app-repos from travis builds
Whenever we build the master branch on the main repo successfully,
travis will now look up all forks with a default branch other than
`master` and send a PR with the changes their way – unless, of
course, there is already a PR pending to merge the latest changes.

Turns out, it isn't easy to tell travis to run a build script only
once in matrix mode. So we are using the 'travis_build_all'-
workaround: this is a script which when run on the first build of
the matrix will observe the other build processes and wait until
they finished successfully. Only then will it exit in 0. In all
other cases it exists 1, allowing us to easily prefix it with "&&"
to make our script only once.

Note: this currently sends PRs with my user account. We probably
want to change that in the future.