Flatten MainMenu Design
Having a nested MainMenu styles means that styles depend on
classes higher in the tree. This isn't only very ineffiecient
in the browser lookup it also makes it much harder to overwrite
specific styles. We are using CSS-Modules however, which means
each class is unique per file, so there is no reason we can't
have them on the top level to improve on both of these points.

This is the change this PR proposes for the MainMenu components.