Set key to entities for latest route
This seems to be the right way to set the key, which on the
front end maps to the Redux reducer. The default is None,
which has no reducer obviously. We might want to think of
some better ways of handling the case where the key does
not match an existing reducer. Perhaps the default should be
something other than None.